Minimize the fields in static_pkgbuild
The purpose of static_pkgbuild is to download sources. This means the
only field that's really needed is sources=(). arch=() is also
needed in case there are arch dependent sources.

makepkg -f should also be used just in case a package with the same name
somehow happens to be built already.

The rest of the required fields serve no purpose so can simply be given
dummy values.