MVP: depend on aur-depends
This integrates aur-depends into rua.

Hovever this is only a MVP, and aur-depends has been inserted into the
middle of the existing system. There's probably more that could be done
to reduce the code count futher, by removing thing's aur-depends can
already handle.

But I'll leave that for after you decide if you want to include this or
not.