Add both solutions and expose problems
The method that works correctly has a quite
ugly code on it.

By the other hand, the iter with `map` is cleaner
but does not work correctly.
This does not show the last item since the zipped
iter ends 1 step before. Not sure if we can simply
add a face entry to the zipped iter to make both
finish at the same time.