Add fakeroot chown override
This should mitigate the problem of a build script
unable to perform prepare() if it asks for chown operations
with e.g. root user.

The problem arises due to the combination of
bubblewrap jail being too strict,
and seemingly fakeroot having a bug.
See https://github.com/vn971/rua/issues/89 for details