Remove allocator overrides
Previously, it was done for binary size IIRC.
It is no longer considered a priority,
and it seems better to simply follow Rust's
standard behavior.

Additionally, `rua` now links against libflate
and libraries like that, so performance becomes
as important as binary size.
(Previously, `rua` didn't do any operations that
take longer than a millisecond or two,
so it made sense to strive for smaller binaries.)